+++ src/app/components/Footer/Footer.tsx
@@ -1,4 +1,5 @@
-import styles from './Footer.module.css';
+import styles from '../../../styles/Footer.module.css';
+
 export default function Footer() {
   return (
     <footer className={styles.footer}>
