+++ src/app/components/Header/Header.tsx
@@ -1,6 +1,6 @@
 import DarkModeToggle from './DarkMode';
 
-import styles from './Header.module.css';
+import styles from '../../../styles/Header.module.css';
 
 export default function Header() {
   return (
